+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.0.10 (WIP)
 
 * Add extra constructor checks before drying values with a `Date` or `RegExp` class name
+* Fix some recursive issues in the `clone()` method
 
 ## 1.0.9 (2018-12-06)